backend/requirements.txt
fastapi
uvicorn
starlette
authlib
httpx
python-dotenv
google-generativeai
pydantic
itsdangerous
email-validator
python-multipart